Added MyFirebaseInstanceIdService to access device registration system & service in Manifest

+++ app/src/main/java/com/openclassrooms/realestatemanager/model/House.kt
@@ -17,7 +17,7 @@ data class House (
         var typeOfHouse: String?,
         var neighborhood: String?,
         var address: String?,
-        var price: Long?,
+        var price: Int?,
         var surface: Int?,
         var numberOfRooms: Int?,
         var numberOfBathRooms: Int?,
